Refactor: move state machine Handle and Worker into separate files
